Move access token related file to their own package
