+++ pkg/access_token/access_token_file.go
@@ -1,4 +1,4 @@
-package storage
+package access_token
 
 import (
 	"fmt"
