+++ pkg/access_token/access_token_handler.go
@@ -1,0 +1,7 @@
+package access_token
+
+type AccessTokenHandler interface {
+	Read() ([]byte, error)
+	Write() error
+	Delete() error
+}
