+++ pkg/access_token/access_token_memory.go
@@ -1,4 +1,4 @@
-package storage
+package access_token
 
 import (
 	"fmt"
